LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select-listbox/select-listbox-home-end-pagedown-pageup.optional-expected.txt



FAIL single: Home/End/PageUp/PageDown should move focus between options and scroll. assert_equals: The first option should be focused at the start of the test. expected Element node <option>0</option> but got Element node <body><select id="single" size="3"><option>0</option><opt...
FAIL multiple: Home/End/PageUp/PageDown should move focus between options and scroll. assert_equals: The first option should be focused at the start of the test. expected Element node <option>0</option> but got Element node <body><select id="single" size="3"><option>0</option><opt...
